* commit 'af0a2a1c34d622de396a45614620cc7854b47011':
  B4:DSP_BT: Add new DDRC target ID configuration parameter
  B4:L1D: Add support for L1 defense New flow
  B4:DSP_BT: DSP cache-cluster init now done from DSP boot loader
  B4:DSP_BT: Add support for DSP MMU configuration

+++ ipc/lib/fsl_ipc_channel.c
@@ -1636,7 +1636,7 @@ int check_validation_fields(void *dsp_bt)
 
 	ipc = (os_het_ipc_t *)(shm.vaddr + (uint32_t)ctrl->ipc - shm.paddr);
 	/* Error prone code check here*/
-	uint32_t phys_addr_sh_ctrl =
+	uint64_t phys_addr_sh_ctrl =
 		((dsp_bt_t *)dsp_bt)->het_sys_map.sh_ctrl_area.phys_addr;
 	os_het_l1d_t *l1_defense = (os_het_l1d_t *)(
 		((dsp_bt_t *)dsp_bt)->sh_ctrl_area.vaddr + ctrl->l1d -
